Main menu branch for cout and c#sama (#111)
* Main menu, settings menu and some of the pause menu

that

* Update Main.unity

* Attempting to fix the enemies not dying.

* Attempt #2 at the fix of the Enemy movement

* Attempt at fixing the object destruction.

* Update Enemy.cs

* Attempt to create the controller for the hiding of the pause menu

* Fixed the empty statemnt with the GetComponent

* Pause in progress

* comments

* Opened screen on pause

* Fixed GameOver Weird Text

* Fixing merge errors

* Update GameController.cs

* Fixed Gameover cycle

* Made the pause menu Escape

* Fixed Main menu

* making health class public

* Fixed enemy paused

* Update MainMenu.cs

* lots o changes

* Fixed enemy movement bug

* left behind

* MainMenu Button

* Level Improvement

* ClickSound

* trying to resolve merging conflict

* bug fixes

* another merging trial

* another one

* Revert "another one"

This reverts commit 3af937ce63b9794264b8c0eadf08f3b29f7e295d.

* more UI stuff

* some bug fixes

* DONE

* fix

+++ BlueProject/Assets/Scripts/Projectile.cs
@@ -8,7 +8,7 @@ namespace Core.Combat
     {
         //set parent when launching the projectile
         [HideInInspector] public GameObject parent;
-        public float damage = 1;
+        public int damage = 1;
         public string hitSFXName;
         public string hitVFXName;
 
@@ -31,7 +31,7 @@ namespace Core.Combat
             }
         }
         */
-        public void TranslateDamage(GameObject target, float attackPower)
+        public void TranslateDamage(GameObject target, int attackPower)
         {
             var health = target.GetComponent<Health>();
             health.TakeDamage(attackPower);
